{42} PlayerFragment´s progressbar now works. It should also now always display the correct information about the currently playing song. The text displayed on the PlayerFragment is also not editable anymore.

+++ app/src/main/java/de/teamawesome/awesomeplayer/playerservice/IPlaybackListener.java
@@ -14,6 +14,6 @@ public interface IPlaybackListener {
 
     void playbackStopped();
 
-    void volumeChanged(int newVolume);
+    void volumeChanged(float newVolume);
 
 }
